project02/requirements.txt
Django
django-filter
djangorestframework
docutils
httpie
